♻️ Refactor: split ChatRoomMessages componet (#63)
