+++ src/components/chat/ChatRoomMessages.tsx
@@ -1,0 +1,53 @@
+import type { ChatRoomMessages as ChatRoomMessagesType } from '@/types/api-response-types/chat-response-types'
+import { cn } from '@/utils'
+
+interface ChatRoomMessagesProps {
+  messages: ChatRoomMessagesType
+}
+
+export default function ChatRoomMessages({ messages }: ChatRoomMessagesProps) {
+  return (
+    <section className="flex flex-col space-y-3">
+      {messages.results.map((result) => {
+        const {
+          content,
+          created_at: createdAt,
+          sender: { nickname, user_uuid: senderId },
+          message_id: id,
+        } = result
+
+        //현재 로그인 한 아이디
+        //추후 실제 아이디와 연결
+        const userId = 'user-2222'
+
+        const isMine = senderId === userId
+
+        const hours = String(createdAt.getHours()).padStart(2, '0')
+        const minutes = String(createdAt.getMinutes()).padStart(2, '0')
+
+        return (
+          <div
+            key={id}
+            className={cn(
+              'flex flex-col space-y-1 p-3',
+              isMine ? 'items-end' : 'items-start'
+            )}
+          >
+            <span className="px-1 text-xs text-gray-600">{nickname}</span>
+            <p
+              className={cn(
+                'rounded-lg px-3 py-2',
+                isMine
+                  ? 'bg-primary-500 rounded-br-xs text-white'
+                  : 'rounded-bl-xs bg-gray-100 text-gray-900'
+              )}
+            >
+              {content}
+            </p>
+            <span className="px-1 text-xs text-gray-500">{`${hours}:${minutes}`}</span>
+          </div>
+        )
+      })}
+    </section>
+  )
+}
